requirements.txt
# Core ML
torch>=2.3.0
torchvision
transformers>=4.46.0
accelerate>=0.33.0
peft>=0.12.0

# vLLM for fast inference
vllm>=0.6.0

# veRL for GRPO training
# Install from source: pip install -e git+https://github.com/volcengine/verl.git
# or: pip install verl

# Multimodal
Pillow>=10.0.0
qwen-vl-utils

# Search tools
requests>=2.31.0
openai>=1.40.0  # DashScope compatible mode

# Dataset
datasets>=2.20.0
huggingface_hub>=0.23.0

# Demo
streamlit>=1.36.0
gradio>=4.40.0

# Utilities
numpy>=1.26.0
tqdm>=4.66.0
python-dotenv>=1.0.0
pydantic>=2.0.0
